structure files, extracting schema and resolvers
