+++ resolvers.js
@@ -1,0 +1,9 @@
+const models = require('./models')
+
+const resolve = {
+  getProducts: async () => (
+    await models.Product.findAll()
+  )
+}
+
+module.exports = resolve;
